auth_provider/requirements.txt
fastapi>=0.100.0
httpx>=0.24.0
python-jose[cryptography]>=3.3.0
uvicorn>=0.22.0